+++ src/main/java/osu_mp3/BeatmapCollection.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 import java.util.ArrayList;
 import java.util.List;
