+++ src/main/java/osu_mp3/BeatmapCollectionDecoder.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 import java.io.*;
 import java.nio.file.Path;
 import java.util.ArrayList;
