+++ src/main/java/osu_mp3/GlobalKeyListener.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 import com.tulskiy.keymaster.common.HotKey;
 import com.tulskiy.keymaster.common.HotKeyListener;
 import com.tulskiy.keymaster.common.MediaKey;
