+++ src/main/java/osu_mp3/MD5Calculator.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
