+++ src/main/java/osu_mp3/Main.java
@@ -1,10 +1,10 @@
-import realm_stuff.RealmReadTester;
+package osu_mp3;
 
 public class Main {
 
     public static void main(final String[] args) throws InterruptedException {
-        //App.launch(App.class, args);
-        RealmReadTester.testRead();
+        App.launch(App.class, args);
+        //RealmReadTester.testRead();
     }
 
     public static void test() {
