+++ src/main/java/osu_mp3/MusicPlayerManager.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 public class MusicPlayerManager {
     
 //    private MusicPlayer musicPlayer = new MusicPlayer();
