+++ src/main/java/osu_mp3/MusicQueueManager.java
@@ -1,4 +1,5 @@
-import java.util.ArrayList;
+package osu_mp3;
+
 import java.util.List;
