+++ src/main/java/osu_mp3/SettingsManager.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
