+++ src/main/java/osu_mp3/SongData.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 public class SongData {
 
     public String songName = "";
