+++ src/main/java/osu_mp3/SongPane.java
@@ -1,3 +1,5 @@
+package osu_mp3;
+
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.layout.Pane;
