+++ src/main/java/realm_database/RealmSchema.kt
@@ -1,4 +1,4 @@
-package realm_stuff
+package realm_database
 
 import io.realm.kotlin.ext.realmListOf
 import io.realm.kotlin.types.*
